tmp_printouts scheduled task: replace the scriptless command with a script & modify scheduled task XML file
This commit adds a new VBScript file and adds the required modifications to the scheduled task XML file, to run the script instead of the command.

Other notes:

* The script will run without showing a window (this is an inherent capability of wscript.exe, and explains my choice of scripting language for such a maintenance task)
* The script will recycle rather than delete files (I plan to add an optional switch for this in the future)
* README.md has been updated to reflect these changes

This commit was made because Microsoft 365 Defender kept flagging the scheduled task action as:
"Use of living-off-the-land binary to run malicious code" & "Suspicious Scheduled Task Process Launched"
(mshta.exe invisibly running Powershell commands)
It also branded the entire sequence of events as a "Multi-stage incident involving Execution & Discovery on multiple endpoints".

Microsoft 365 Defender should not flag the new script and associated scheduled task as malicious.

Side-note: admittedly the method previously employed was a bit of a hack to get the task to run without showing a Powershell window, but there was no malintent.