* Reorganise helper-scripts by subdividing them into two new subfolders, setup-scripts and maintenance-scripts * tmp_printouts: replace recycleFiles.vbs with recycleFiles.ps1, add silent.vbs & modify scheduled task XML file accordingly * Update README.md to reflect directory and file name changes
The major change in this commit is to replace the VBScript with a Powershell equivalent as this is generally the way forward. However, at least for now — VBS in the form of silent.vbs — still needs to be employed in order to run the task invisibly.

See also previous commit c85b5489a5c699f9ac24e01131f817d1bd104c8f